requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
groq==0.11.0
PyMuPDF==1.24.10
python-docx==1.1.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers 